made changes for pull request #97
