+++ src/spas/apps/jwadmin/src/app/app.component.html
@@ -1,0 +1,1 @@
+<member-management></member-management>
