+++ src/spas/apps/jwadmin/src/app/app.module.ts
@@ -4,12 +4,14 @@ import { AngularFireModule } from '@angular/fire';
 
 import { AppComponent } from './app.component';
 import { environment } from '../environments/environment';
+import { MemberManagementFeatureModule } from '@nimb3s/member-management/feature'
 
 @NgModule({
   declarations: [AppComponent,],
   imports: [
     BrowserModule,
-    AngularFireModule.initializeApp(environment.firebase, 'dev-jwadmin')
+    AngularFireModule.initializeApp(environment.firebase, 'dev-jwadmin'),
+    MemberManagementFeatureModule
   ],
   providers: [],
   bootstrap: [AppComponent]
