+++ src/spas/apps/nimb3s/src/app/app.component.html
@@ -1,3 +1,3 @@
 <h1>Welcome to nimb3s!</h1>
 
-<jwadmin-member-management></jwadmin-member-management>
+
