+++ src/spas/libs/member-management/domain/src/index.ts
@@ -1,1 +1,3 @@
 export * from './lib/member-management-domain';
+export * from './lib/mocks/mock-users';
+export * from './lib/models/user-profile';
