+++ src/spas/libs/member-management/domain/src/lib/models/user-profile.ts
@@ -1,0 +1,11 @@
+export class User {
+  fullName: string;
+  emailAddress: string;
+  personalPhoneNumber: number;
+  homeAddress: string;
+  emergencyContactName: string;
+  emergencyContactPhone: number;
+  signUpDate: string;
+  congregationName: string;
+  groupNumber: number;
+}
