+++ src/spas/libs/member-management/domain/src/lib/user-service-interface.ts
@@ -1,0 +1,6 @@
+import { Observable } from 'rxjs';
+import { User } from './models/user-profile';
+
+export interface UserInterface {
+  getUsers: Observable<User[]>;
+ }
