--- src/spas/libs/member-management/feature/src/lib/member-management/user-profile.ts
@@ -1,11 +1,0 @@
-export class User {
-  fullName: string;
-  emailAddress: string;
-  personalPhoneNumber: number;
-  homeAddress: string;
-  emergencyContactName: string;
-  emergencyContactPhone: number;
-  signUpDate: string;
-  congregationName: string;
-  groupNumber: number;
-}
