+++ src/spas/libs/member-management/feature/src/lib/member-management/users/users.component.html
@@ -1,6 +1,6 @@
 <h2>Users</h2>
 <ul>
-  <li *ngFor="let user of users" (click)="onSelect(user)">
+  <li *ngFor="let user of users | async" (click)="onSelect(user)">
     {{user.fullName}}
   </li>
 </ul>
