+++ src/spas/libs/member-management/feature/tsconfig.json
@@ -3,5 +3,5 @@
   "compilerOptions": {
     "types": ["node", "jest"]
   },
-  "include": ["**/*.ts"]
+  "include": ["**/*.ts", "../domain/src/lib/mocks/mock-users.ts", "../domain/src/lib/models/user-profile.ts", "../domain/src/lib/user.service.ts", "../domain/src/lib/user.service.spec.ts"]
 }
